requirements.txt
pandas
numpy
requests
diskcache
joblib
plotly
flask
python-dotenv
flipside
prophet
json
scikit-learn
pyngrok
ssl
sqlite3
eth-account
web3
polars
pyarrow
